Move dossier template indexers into package.
Move the dossier template indexers into the dossiertemplate package.
