+++ opengever/dossier/dossiertemplate/configure.zcml
@@ -3,6 +3,16 @@
     xmlns:browser="http://namespaces.zope.org/browser"
     i18n_domain="opengever.dossier">
 
+  <adapter
+      factory=".indexers.DossierTemplateSubjectIndexer"
+      name="Subject"
+      />
+
+  <adapter
+      factory=".indexers.DossierTemplateSearchableTextExtender"
+      name="IDossierTemplate"
+      />
+
   <browser:page
       for="opengever.repository.interfaces.IRepositoryFolder"
       name="dossier_with_template"
